@@ -1,18 +1,20 @@
 source 'https://rubygems.org'
 
-ruby '2.7.7'
+ruby '3.3.6'
 
-gem 'rails', '6.1.7.3'
+gem 'rails', '7.2.3'
 gem 'rake'
 gem 'bootsnap','>= 1.1.0', require: false
 
 gem 'addressable'
+gem 'anyway_config'
 gem 'aws-sdk-s3'
 gem 'bourbon', '~> 4.0.2'
 gem 'bcrypt', '~> 3.1.7'
 gem 'clearance', '~> 2.5.0'
 gem 'clearance-deprecated_password_strategies'
 gem 'coffee-rails'
+gem 'csv'
 gem 'flutie'
 gem 'formtastic'
 gem 'friendly_id', '~> 5.5.0'
@@ -24,12 +26,11 @@ gem 'jbuilder'
 gem 'jquery-rails', '~> 4.1.1'
 gem 'jquery-ui-rails', '~> 3.0.1'
 gem 'magnific-popup-rails'
-gem 'mail', '~> 2.7.1' # TODO see if this is still required in Ruby 3
 gem 'nio4r', '>= 2.7.3'
-gem 'nokogiri', '~> 1.14.3'
+gem 'nokogiri', '~> 1.18.9'
 gem 'ruby-openai'
 gem 'pg', '~> 1.4.4'
-gem 'puma', '< 7'
+gem 'puma'
 gem 'rack-attack'
 gem 'rack-cors'
 gem 'rack-ssl-enforcer'
@@ -40,10 +41,11 @@ gem 'sassc-rails', '~> 2.1'
 gem 'shrine', '~> 3'
 gem 'shrine-tus', '~> 2.1', require: false
 gem 'simple_form', '~> 5.1'
+gem 'sprockets-rails', require: "sprockets/railtie"
 gem 'stimulus-rails'
 gem 'sucker_punch', '~> 3.1'
 gem 'terser'
-gem 'textacular', '~> 5.5.1'
+gem 'textacular', '~> 5.7'
 gem 'turbo-rails'
 gem 'tus-server', '~> 2.3', require: false
 gem 'uppy-s3_multipart', '~> 1.2'
@@ -62,28 +64,28 @@ group :development, :test do
   gem 'byebug'
   gem 'pry'
   gem 'pry-nav'
-  gem 'rspec-rails', '~> 6.0.1'
+  gem 'rspec-rails', '~> 7.0.2'
   gem 'sham_rack'
 end
 
 group :test do
   gem 'bourne'
-  gem 'capybara', '~> 3'
+  gem 'capybara'
   gem 'capybara-screenshot', '~> 1.0'
   gem 'database_cleaner'
   gem 'email_spec'
-  gem 'factory_girl_rails'
+  gem 'factory_bot_rails'
   gem 'faker'
   gem 'launchy'
   gem 'rails-controller-testing'
+  gem 'selenium-webdriver'
   gem 'shoulda-matchers', '~> 3'
   gem 'timecop'
   gem 'turnip'
-  gem 'webdrivers'
 end
 
 group :staging, :production do
-  gem 'newrelic_rpm', '~> 7.2'
+  gem 'newrelic_rpm'
   gem 'rails_12factor'
   gem 'sprockets-redirect'
 end
