@@ -16,6 +16,7 @@ EOT
   has_many :roles
   has_many :users, :through => :roles
   has_many :projects
+  has_many :projects_pending_moderation, -> { joins(:project_moderation).merge(ProjectModeration.pending) }, class_name: "Project"
   has_many :winning_projects, -> { where.not(funded_on: nil).order(funded_on: :desc) }, class_name: "Project"
   has_many :invitations
 
@@ -63,13 +64,13 @@ EOT
 
   # Scope can be :active or :all
   def self.select_data(scope = :active, include_any: true)
-    countries = include_any ? [OpenStruct.new(:name => "Any Chapter", :chapters => where(name: ANY_CHAPTER_NAME))] : []
+    countries = include_any ? [Collection.new(:name => "Any Chapter", :chapters => where(name: ANY_CHAPTER_NAME))] : []
 
     selection = scope == :all ? visitable : active.visitable
 
     selection.sort_by(&CountrySortCriteria.new(COUNTRY_PRIORITY)).each do |chapter|
       if countries.last.try(:name) != chapter.country
-        countries.push OpenStruct.new(:name => chapter.country, :chapters => [])
+        countries.push Collection.new(:name => chapter.country, :chapters => [])
       end
 
       countries.last.chapters.push chapter
@@ -142,6 +143,16 @@ EOT
     if account = twitter_url.to_s.split("/").compact.last
       account = "@#{account}" unless account.match(/^@/)
       account
+    end
+  end
+
+  class Collection
+    attr_accessor :name
+    attr_accessor :chapters
+
+    def initialize(name:, chapters:)
+      @name = name
+      @chapters = chapters
     end
   end
 end
