@@ -26,9 +26,9 @@ describe ProjectsController do
   context 'viewing the index with a search term' do
     render_views
 
-    let(:user) { FactoryGirl.create(:user) }
-    let(:project) { FactoryGirl.create(:project, title: "Find Me") }
-    let!(:missing_project) { FactoryGirl.create(:project, chapter: project.chapter) }
+    let(:user) { FactoryBot.create(:user) }
+    let(:project) { FactoryBot.create(:project, title: "Find Me") }
+    let!(:missing_project) { FactoryBot.create(:project, chapter: project.chapter) }
 
     before do
       sign_in_as user
@@ -44,8 +44,8 @@ describe ProjectsController do
   end
 
   context 'attempting to delete a project as a trustee who is not the dean or an admin' do
-    let(:user) { FactoryGirl.create(:user) }
-    let(:project) { FactoryGirl.create(:project) }
+    let(:user) { FactoryBot.create(:user) }
+    let(:project) { FactoryBot.create(:project) }
     before do
       sign_in_as user
       delete :destroy, params: { id: project }
@@ -54,9 +54,9 @@ describe ProjectsController do
   end
 
   context 'viewing the index without a chapter' do
-    let(:chapter) { FactoryGirl.create(:chapter) }
-    let(:user) { FactoryGirl.create(:user) }
-    let!(:role) { FactoryGirl.create(:role, :user => user, :chapter => chapter) }
+    let(:chapter) { FactoryBot.create(:chapter) }
+    let(:user) { FactoryBot.create(:user) }
+    let!(:role) { FactoryBot.create(:role, :user => user, :chapter => chapter) }
 
     before do
       sign_in_as user
@@ -67,10 +67,10 @@ describe ProjectsController do
   end
 
   context 'downloading the csv report' do
-    let!(:project)  { FactoryGirl.create :project }
-    let!(:user)     { FactoryGirl.create :user }
-    let!(:chapter)  { FactoryGirl.create :chapter }
-    let!(:role)     { FactoryGirl.create :role, :user => user, :chapter => chapter }
+    let!(:project)  { FactoryBot.create :project }
+    let!(:user)     { FactoryBot.create :user }
+    let!(:chapter)  { FactoryBot.create :chapter }
+    let!(:role)     { FactoryBot.create :role, :user => user, :chapter => chapter }
 
     before do
       sign_in_as user
@@ -83,9 +83,9 @@ describe ProjectsController do
   context "viewing a private project page" do
     render_views
 
-    let!(:project) { FactoryGirl.create(:project) }
-    let!(:trustee) { FactoryGirl.create(:user) }
-    let!(:admin)   { FactoryGirl.create(:admin) }
+    let!(:project) { FactoryBot.create(:project) }
+    let!(:trustee) { FactoryBot.create(:user) }
+    let!(:admin)   { FactoryBot.create(:admin) }
 
     context "while not logged in" do
       before do
@@ -117,6 +117,104 @@ describe ProjectsController do
   end
 
   context "creating a project submission" do
+    let(:fake_mailer) { FakeMailer.new }
+    let(:chapter) { FactoryBot.create(:chapter) }
+    let(:valid_project_params) do
+      {
+        chapter_id: chapter.id,
+        name: "Test User",
+        title: "Test Project",
+        email: "test@example.com",
+        about_me: "About me",
+        about_project: "About project",
+        use_for_money: "Use for money"
+      }
+    end
+
+    before do
+      Project.any_instance.stubs(:mailer).returns(fake_mailer)
+    end
+
+    context "that is valid" do
+      it "creates a project and sends an email" do
+        expect {
+          post :create, params: { project: valid_project_params }
+        }.to change(Project, :count).by(1)
+
+        expect(response).to redirect_to(success_submissions_path(chapter: chapter))
+        expect(fake_mailer).to have_delivered_email(:new_application)
+      end
+    end
+
+    context "that is suspected of being spam" do
+      it "creates a project but does not send an email" do
+        Project.any_instance.stubs(:project_moderation).returns(ProjectModeration.new(status: :suspected))
+
+        expect {
+          post :create, params: { project: valid_project_params }
+        }.to change(Project, :count).by(1)
+
+        expect(response).to redirect_to(success_submissions_path(chapter: chapter))
+        expect(Project.last.not_pending_moderation?).to eq(false)
+        expect(fake_mailer).to_not have_delivered_email(:new_application)
+      end
+    end
+
+    context "with client_metadata" do
+
+      it "sanitizes malicious client_metadata fields" do
+        malicious_metadata = {
+          time_on_page_ms: "not_a_number",
+          timezone: "x" * 1000, # Very long string
+          screen_resolution: "1920x1080" * 100, # Very long string
+          form_interactions_count: "5",
+          keystroke_count: 42,
+          paste_count: "invalid"
+        }.to_json
+
+        post :create, params: {
+          project: valid_project_params,
+          client_metadata: malicious_metadata
+        }
+
+        expect(response).to redirect_to(success_submissions_path(chapter: chapter))
+
+        project = Project.last
+        expect(project.metadata["time_on_page_ms"]).to be_nil # Invalid integer
+        expect(project.metadata["timezone"].length).to eq(50) # Truncated
+        expect(project.metadata["screen_resolution"].length).to eq(20) # Truncated
+        expect(project.metadata["form_interactions_count"]).to eq(5) # Valid integer from string
+        expect(project.metadata["keystroke_count"]).to eq(42) # Valid integer
+        expect(project.metadata["paste_count"]).to be_nil # Invalid integer
+      end
+
+      it "processes valid client_metadata correctly" do
+        valid_metadata = {
+          time_on_page_ms: 5000,
+          timezone: "America/New_York",
+          screen_resolution: "1920x1080",
+          form_interactions_count: 3,
+          keystroke_count: 42,
+          paste_count: 1
+        }.to_json
+
+        post :create, params: {
+          project: valid_project_params,
+          client_metadata: valid_metadata
+        }
+
+        expect(response).to redirect_to(success_submissions_path(chapter: chapter))
+
+        project = Project.last
+        expect(project.metadata["time_on_page_ms"]).to eq(5000)
+        expect(project.metadata["timezone"]).to eq("America/New_York")
+        expect(project.metadata["screen_resolution"]).to eq("1920x1080")
+        expect(project.metadata["form_interactions_count"]).to eq(3)
+        expect(project.metadata["keystroke_count"]).to eq(42)
+        expect(project.metadata["paste_count"]).to eq(1)
+      end
+    end
+
     context "when incomplete but with an image" do
       render_views
 
@@ -127,13 +225,14 @@ describe ProjectsController do
       it { is_expected.to respond_with(:success) }
       it { is_expected.to render_template("new") }
     end
+
   end
 
   context "the project submission form" do
     let(:mode) { nil }
 
     context "with a chapter" do
-      let(:chapter) { FactoryGirl.create(:chapter) }
+      let(:chapter) { FactoryBot.create(:chapter) }
 
       before do
         get :new, params: { chapter: chapter.slug, mode: mode }
@@ -180,9 +279,9 @@ describe ProjectsController do
   end
 
   describe "#hide" do
-    let(:user) { FactoryGirl.create(:user) }
+    let(:user) { FactoryBot.create(:user) }
     let(:reason) { Faker::Company.bs }
-    let(:project) { FactoryGirl.create(:project) }
+    let(:project) { FactoryBot.create(:project) }
     before do
       sign_in_as user
     end
@@ -238,9 +337,9 @@ describe ProjectsController do
   end
 
   describe "#unhide" do
-    let(:user) { FactoryGirl.create(:user) }
+    let(:user) { FactoryBot.create(:user) }
     let(:reason) { Faker::Company.bs }
-    let(:project) { FactoryGirl.create(:project) }
+    let(:project) { FactoryBot.create(:project) }
     before do
       sign_in_as user
       project.hide!(reason, user)
