@@ -1,14 +1,14 @@
 require 'spec_helper'
 
 describe WinnersController do
-  let(:other_chapter) { FactoryGirl.create(:chapter) }
+  let(:other_chapter) { FactoryBot.create(:chapter) }
 
   before do
     sign_in_as(user) if defined?(user)
   end
 
   context "user is not logged in" do
-    let(:project) { FactoryGirl.create(:project) }
+    let(:project) { FactoryBot.create(:project) }
 
     it "redirects when trying to view the edit page" do
       get :edit, params: { chapter_id: project.chapter_id, project_id: project.id }
@@ -20,10 +20,10 @@ describe WinnersController do
   end
 
   context "a dean is logged in" do
-    let(:user) { FactoryGirl.create(:user_with_dean_role) }
+    let(:user) { FactoryBot.create(:user_with_dean_role) }
 
     context "a project is in a real chapter" do
-      let(:project) { FactoryGirl.create(:project) }
+      let(:project) { FactoryBot.create(:project) }
 
       it "cannot set the winner for another chapter" do
         post :create, params: { project_id: project.id, chapter_id: other_chapter.id }
@@ -34,7 +34,7 @@ describe WinnersController do
     end
 
     context "a project is in the Any chapter" do
-      let(:project) { FactoryGirl.create(:project, chapter: Chapter.find_by_name('Any')) }
+      let(:project) { FactoryBot.create(:project, chapter: Chapter.find_by_name('Any')) }
 
       it "cannot set the winner for another chapter" do
         post :create, params: { project_id: project.id, chapter_id: other_chapter.id }
@@ -46,10 +46,10 @@ describe WinnersController do
   end
 
   context "an admin is logged in" do
-    let(:user) { FactoryGirl.create(:admin) }
+    let(:user) { FactoryBot.create(:admin) }
 
     context "a project is in a real chapter" do
-      let(:project) { FactoryGirl.create(:project) }
+      let(:project) { FactoryBot.create(:project) }
 
       it "can set the winner for another chapter" do
         post :create, params: { project_id: project.id, chapter_id: other_chapter.id }
@@ -60,7 +60,7 @@ describe WinnersController do
     end
 
     context "a project is in the Any chapter" do
-      let(:project) { FactoryGirl.create(:project, chapter: Chapter.find_by_name('Any')) }
+      let(:project) { FactoryBot.create(:project, chapter: Chapter.find_by_name('Any')) }
 
       it "can set the winner for another chapter" do
         post :create, params: { project_id: project.id, chapter_id: other_chapter.id }
