@@ -1,5 +1,5 @@
 step 'I promote the trustee to admin' do
-  @current_chapter = FactoryGirl.create(:chapter)
+  @current_chapter = FactoryBot.create(:chapter)
   visit chapter_projects_path(@current_chapter)
   click_link("View all Users")
   page.find(:css, "tr[data-user-id='#{@trustee.id}'].non_admin td.promote-demote-admin a.promote-user").click
