@@ -26,7 +26,7 @@ step 'I should see the new chapter form with errors' do
 end
 
 step 'there is a chapter in the system' do
-  @current_chapter = FactoryGirl.create(:chapter, :rss_feed_url => Rails.root.join('spec', 'support', 'feed.xml').to_s)
+  @current_chapter = FactoryBot.create(:chapter, :rss_feed_url => Rails.root.join('spec', 'support', 'feed.xml').to_s)
 end
 
 step 'the chapter has trustees hidden' do
@@ -34,7 +34,7 @@ step 'the chapter has trustees hidden' do
 end
 
 step 'there is an inactive chapter in the system' do
-  @current_chapter = FactoryGirl.create(:inactive_chapter)
+  @current_chapter = FactoryBot.create(:inactive_chapter)
 end
 
 step 'I go to the chapter page' do
@@ -135,7 +135,7 @@ step 'I should see a permissions error' do
 end
 
 step 'there are 5 chapters' do
-  @chapters = (1..5).map{ FactoryGirl.create(:chapter) }
+  @chapters = (1..5).map{ FactoryBot.create(:chapter) }
 end
 
 step "those projects' chapters are in 4 countries total" do
