@@ -1,7 +1,7 @@
 step 'there is a project with :count photo(s)' do |count|
-  @chapter = FactoryGirl.create(:chapter)
-  @project = FactoryGirl.create(:project, :chapter => @chapter, :funded_on => 1.days.ago)
-  FactoryGirl.create_list(:photo, count.to_i,  :project => @project)
+  @chapter = FactoryBot.create(:chapter)
+  @project = FactoryBot.create(:project, :chapter => @chapter, :funded_on => 1.days.ago)
+  FactoryBot.create_list(:photo, count.to_i,  :project => @project)
 end
 
 step 'I view the project' do
