@@ -26,7 +26,7 @@ step 'I should not see promotion links' do
 end
 
 step 'I am a trustee for another chapter as well' do
-  @other_current_chapter = FactoryGirl.create(:chapter)
-  FactoryGirl.create(:role, :chapter => @other_current_chapter,
+  @other_current_chapter = FactoryBot.create(:chapter)
+  FactoryBot.create(:role, :chapter => @other_current_chapter,
                 :user => @current_user)
 end
