@@ -34,33 +34,33 @@ step 'none of the projects should be shortlisted' do
 end
 
 step 'someone has voted for a project in another chapter' do
-  FactoryGirl.create(:vote)
+  FactoryBot.create(:vote)
 end
 
 step ':count people/person have/has voted on a/another project in my chapter' do |count|
   @finalist_projects ||= []
-  project = FactoryGirl.create(:project, :chapter => @current_chapter, :created_at => count.to_i.days.ago)
+  project = FactoryBot.create(:project, :chapter => @current_chapter, :created_at => count.to_i.days.ago)
   @finalist_projects << [project, count] if count.to_i > 0
   count.to_i.times do |x|
-    role = FactoryGirl.create(:role, :chapter => @current_chapter)
-    vote = FactoryGirl.create(:vote, :user => role.user, :project => project, :chapter => @current_chapter, :created_at => x.days.ago)
+    role = FactoryBot.create(:role, :chapter => @current_chapter)
+    vote = FactoryBot.create(:vote, :user => role.user, :project => project, :chapter => @current_chapter, :created_at => x.days.ago)
   end
 end
 
 step 'there are some projects for this month with votes' do
-  role = FactoryGirl.create(:role, :chapter => @current_chapter)
+  role = FactoryBot.create(:role, :chapter => @current_chapter)
   3.times do
-    project = FactoryGirl.create(:project, :chapter => @current_chapter)
-    FactoryGirl.create(:vote, :user => role.user, :project => project, :chapter => @current_chapter)
+    project = FactoryBot.create(:project, :chapter => @current_chapter)
+    FactoryBot.create(:vote, :user => role.user, :project => project, :chapter => @current_chapter)
   end
 end
 
 step 'there are some projects in the "Any" chapter for this month with votes' do
   @any_chapter = Chapter.where(:name => "Any").first
-  role = FactoryGirl.create(:role, :chapter => @current_chapter)
+  role = FactoryBot.create(:role, :chapter => @current_chapter)
   3.times do
-    project = FactoryGirl.create(:project, :chapter => @any_chapter)
-    FactoryGirl.create(:vote, :user => role.user, :project => project, :chapter => @current_chapter)
+    project = FactoryBot.create(:project, :chapter => @any_chapter)
+    FactoryBot.create(:vote, :user => role.user, :project => project, :chapter => @current_chapter)
   end
 end
 
