@@ -2,12 +2,12 @@ require "spec_helper"
 
 describe InvitationsHelper do
 
-  let!(:role) { FactoryGirl.create(:role, :name => 'dean') }
+  let!(:role) { FactoryBot.create(:role, :name => 'dean') }
   let!(:user) { role.user }
   let!(:chapter) { role.chapter }
-  let!(:another_role) { FactoryGirl.create(:role, :name => 'dean', :user => user) }
+  let!(:another_role) { FactoryBot.create(:role, :name => 'dean', :user => user) }
   let!(:another_chapter) { another_role.chapter }
-  let!(:yet_another_role) { FactoryGirl.create(:role, :name => 'dean', :user => user) }
+  let!(:yet_another_role) { FactoryBot.create(:role, :name => 'dean', :user => user) }
   let!(:yet_another_chapter) { yet_another_role.chapter }
   let!(:invitation) { Invitation.new }
 
