@@ -3,8 +3,8 @@ require 'date'
 
 describe ProjectNavHelper do
   describe '#project_winning_siblings' do
-    let(:chapter) { FactoryGirl.create :chapter, :name => 'Test chapter' }
-    let!(:project) { FactoryGirl.create :winning_project, :chapter => chapter }
+    let(:chapter) { FactoryBot.create :chapter, :name => 'Test chapter' }
+    let!(:project) { FactoryBot.create :winning_project, :chapter => chapter }
 
     context 'for a single winning project' do
       it 'returns no siblings' do
@@ -12,21 +12,21 @@ describe ProjectNavHelper do
       end
 
       it 'does not include not winning projects as siblings' do
-        not_winning_projects = FactoryGirl.create_list :project, 2, :chapter => chapter
+        not_winning_projects = FactoryBot.create_list :project, 2, :chapter => chapter
         expect(helper.project_winning_siblings(project).compact).to be_empty
       end
     end
 
     context 'for a single not winning project' do
       it 'returns no siblings' do
-        not_winning_project = FactoryGirl.create :project, :chapter => chapter
+        not_winning_project = FactoryBot.create :project, :chapter => chapter
         expect(helper.project_winning_siblings(not_winning_project).compact).to be_empty
       end
     end
 
     context 'with winning projects' do
-      let!(:next_project)     { FactoryGirl.create :project, :chapter => chapter, :funded_on => project.funded_on - 1.day }
-      let!(:previous_project) { FactoryGirl.create :project, :chapter => chapter, :funded_on => project.funded_on + 1.day }
+      let!(:next_project)     { FactoryBot.create :project, :chapter => chapter, :funded_on => project.funded_on - 1.day }
+      let!(:previous_project) { FactoryBot.create :project, :chapter => chapter, :funded_on => project.funded_on + 1.day }
 
       it 'returns Next sibling for first project' do
         expect(helper.project_winning_siblings(previous_project).last).to eq(project)
