@@ -1,8 +1,8 @@
 require 'spec_helper'
 
 describe 'shared/_navigation' do
-  let!(:chapter) { FactoryGirl.create(:chapter) }
-  let!(:inactive_chapter) { FactoryGirl.create(:inactive_chapter) }
+  let!(:chapter) { FactoryBot.create(:chapter) }
+  let!(:inactive_chapter) { FactoryBot.create(:inactive_chapter) }
 
   it 'only displays active chapters in the chapter list' do
     view.stubs(:signed_in?).returns(false)
