@@ -4,4 +4,4 @@
 //
 // By default, this file is loaded for server-side rendering.
 // It should require your components and any dependencies.
-//=require ./awesome_react
+//= require ./awesome_react
